Remove GuidGenerator in favor of free function
Old way:

  GuidGenerator gen;
  auto g = gen.newGuid();

New way:

  auto g = xg::newGuid();

The new way is better, but not if you're on android. The GuidGenerator
indirection existed to make life better on android since you need some
context (JNIEnv *); however, the new strategy is to let android be weird
and optimize the API everywhere else. This is how you do it on android
now:

  xg::initJni(jniEnv); // only do this once
  auto g = xg::newGuid();