Add delete permissions to the include files
Currently the install files are copied with permissions that do not allow a normal user to delete them. This causes a problem when the code is cloned and built in a folder as part of a nightly build. We are unable to remove the folder ready for the next night without using sudo.
Would you consider accepting my change as other people might benefit from it.